Forget password done.

+++ src/main/java/com/Kamol/SalonSync/security/jwt/JwtUtils.java
@@ -21,7 +21,7 @@ public class JwtUtils {
     private String jwtSecret;
 
     @Value("${SalonSync.app.jwtExpirationMs}")
-    private int jwtExpirationMs;
+    private Long jwtExpirationMs;
 
     public String generateJwtToken(Authentication authentication) {
